feat: implement basic login screen actions
