+++ client/app/src/main/java/com/example/criengine/Activities/MainActivity.java
@@ -29,11 +29,6 @@ import java.util.List;
 
 public class MainActivity extends AppCompatActivity {
 
-    private final boolean CONFIG_LOGIN_AUTOMATICALLY = true;
-    private final String CONFIG_AUTO_LOGIN_EMAIL = "user2@email.com";
-    private final String CONFIG_AUTO_LOGIN_PASSWORD = "password";
-
-
     private Button myProfileActivityButton;
     private Button myBooksActivityButton;
     private Button requestActivityButton;
@@ -47,6 +42,9 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
+        dbw = DatabaseWrapper.getWrapper();
+        assert dbw != null;
 
 
         requestActivityButton = findViewById(R.id.requested_books_activity);
@@ -99,60 +97,6 @@ public class MainActivity extends AppCompatActivity {
                 v.getContext().startActivity(intent);
             }
         });
-
-        //begin sign in
-        // accounts: user1@email.com:password
-        //           user2@email.com:password
-        final FirebaseAuth mAuth = FirebaseAuth.getInstance();
-        if (mAuth.getCurrentUser() == null || CONFIG_LOGIN_AUTOMATICALLY) {
-            if (CONFIG_LOGIN_AUTOMATICALLY) {
-                // Sign in with given username and password
-                mAuth.signInWithEmailAndPassword(CONFIG_AUTO_LOGIN_EMAIL, CONFIG_AUTO_LOGIN_PASSWORD)
-                    .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
-                        @Override
-                        public void onComplete(@NonNull Task<AuthResult> task) {
-                            if (task.isSuccessful()) {
-                                // Sign in success, update UI with the signed-in user's information
-                                final FirebaseUser user = mAuth.getCurrentUser();
-                                dbw = new DatabaseWrapper(user);
-//                                dbw.getProfile(user.getUid()).addOnSuccessListener(
-//                                        new OnSuccessListener<Profile>() {
-//                                            @Override
-//                                            public void onSuccess(Profile profile) {
-//                                                // Add a profile if none existed for this user
-//                                                dbw.addProfile(profile);
-//                                            }
-//                                        }
-//                                );
-                            } else {
-                                // If sign in fails, display a message to the user.
-                                Log.w("TAG", "signInWithEmail:failure", task.getException());
-                                // ...
-                            }
-
-                            // ...
-                        }
-                });
-        } else {
-                // Create a sign in activity
-                List<AuthUI.IdpConfig> providers = Arrays.asList(
-                        new AuthUI.IdpConfig.EmailBuilder().build());
-                startActivityForResult(
-                        AuthUI.getInstance()
-                                .createSignInIntentBuilder()
-                                .setAvailableProviders(providers)
-                                .setIsSmartLockEnabled(false)
-//                        .setLogo(R.drawable.my_great_logo)
-//                        .setTheme(R.style.MySuperAppTheme)
-                                .build(),
-                        1);
-            }
-        } else {
-            if (dbw == null) {
-                dbw = new DatabaseWrapper(mAuth.getCurrentUser());
-            }
-        }
-
     }
 
     @Override
